Logic for transaction package added

+++ widgetBe/src/main/java/home/officers/backend/widgetbe/transactions/model/dto/TransactionsResponseDto.java
@@ -6,6 +6,6 @@ public record TransactionsResponseDto(
         CustomerDto customer,
         int cashTransactions,
         int onlineTransactions,
-        long carbonFootprint
+        double carbonFootprint
 ) {
 }
